+++ mayan/apps/appearance/links.py
@@ -7,13 +7,17 @@ from mayan.apps.navigation.classes import Link
 from mayan.apps.navigation.utils import factory_condition_queryset_access
 
 from .icons import (
-    icon_theme_create, icon_theme_delete, icon_theme_edit, icon_theme_list,
-    icon_theme_setup, icon_user_theme_settings_detail,
+    icon_ajax_refresh, icon_theme_create, icon_theme_delete, icon_theme_edit,
+    icon_theme_list, icon_theme_setup, icon_user_theme_settings_detail,
     icon_user_theme_settings_edit
 )
 from .permissions import (
     permission_theme_create, permission_theme_delete, permission_theme_edit,
     permission_theme_view
+)
+
+link_ajax_refresh = Link(
+    icon=icon_ajax_refresh, html_extra_classes='appearance-link-ajax-refresh'
 )
 
 link_user_theme_settings_detail = Link(